(*) moved scripts and standalone tools into tools/ directory
Script are different from "apps": they are interpreted (with Bash or Python),
while apps are compiled. Standalone tools do not depend on the make system.
It makes sense to keep them in a separate place, such as the tools/ directory.
The files in apps/miscell directory should in fact be in tools/, because they
are all scripts.

In general, tools/ directory should contain:
 - converter scripts;
 - simple intial data generators;
 - SLURM submission scripts;
 - visualization scripts.