(*) boolean parameters shouldn't be initialized to string
When defining new boolean parameters with a macro, e.g.

  DECLARE_PARAM(bool,equal_mass,true)

initialize it to "true" or "false" rather than "yes" and "no",
because underneath they are just variables of type bool:

  bool equal_mass = true;

 - it doesn't make sense to initialize them with strings "yes"
or "no".

On the other hand, in parameter files the "yes/no" syntax is
acceptable, because the parser reads "yes" or "no" to begin with.
I've also now added possibility for the parser to understand
"true", for example this syntax in a parfile would work:

  equal_mass = "true"
  equal_mass = 'true'
  equal_mass = true
  equal_mass = yes (also understands 'yes' and "yes")

Every other string except for these six options in a parameter
file is treated as "false" and the false value is assigned to
that boolean parameter.