Drag force is started to add
Damping (do_drag in param) is controlled relaxation of particle.
The value of drag_coeff is determined by dt which is used in the initial
file, but DO NOT rescale by time step. Relaxation steps (relax_steps in param)
need enough numbers to stablize. Need to check which values of parameters should
be applied to code.