(*) renamed parameter to sph_update_uniform_h
The "variable_smoothinglength" could also be misinterpreted as
variable in space, which is not what we want.
Also, removed this parameter from Sedov test, because in that test
an internal region develops very low density and correspondingly
too high smoothing length to be good for properly resolving the
shock itself.