(+) ability to specify several external potentials
There is now a parser in the eforce.h function select():
one can specify several boundary, bulk or drag forces, using
a comma-separated list of external potentials. For example,
to include bulk gravity, walls in y- and z-directions, and an
airfoil:

  external_force_type = "walls:yz,gravity,airfoil"

The 'walls:???' allows to specify where the walls should be located.