(*) in inputDataHDF5(): use input_file_prefix instead of inputfile
Changed semantics of the inputDataHDF5 to accept only the file prefix
in a second argument. This makes function signature more consistent,
as the third argument also is a prefix (output_file_prefix). It also
allows in the future to develop a handler for a case where the input
is describing a set of HDF5 snapshots, i.e. "sim_00000.h5part" etc.

The function 'inputDataHDF5' is only called from 'read_bodies', which
in turn is called from both drivers and tests.