(!) fix for the "Assertion (i<20) failed"
In the spherical density support potential and acceleration, if a
particle went out of the R=1 sphere, the code would crash, due to
NaNs from the pow() function, which apparently cannot handle zeros.

Additionally, if a particle escapes the density-supporting well,
it accelerates away, which may lead to other issues. To fix this,
added spherical wall potential *outside* R to bounce such particles
back to the well.